Rewrite <%= to not trigger Lint/Void cop
Instances of <%= where the contents do not call a method to generate the
output will trigger the Lint/Void cop since rubocop does not see
anything handle the result of the code.

A simple fix is to replace this with an imaginary method erb which
luckily is also 3 characters long, keeping our columns numbers intact.

Resolves #16